refactor: Replaced IsBlocked with IsRestricted, renamed linked methods and their logic
