+++ Streetcode/Streetcode.BLL/DTO/Comments/CommentDTO.cs
@@ -1,5 +1,6 @@
 ﻿using Streetcode.BLL.DTO.Users;
 using Streetcode.DAL.Entities.Comments;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Streetcode.BLL.DTO.Comments;
 
@@ -14,10 +15,12 @@ public class CommentDTO
     public int StreetcodeId { get; set; }
     public int? ParentCommentId { get; set; }
     public bool IsDeleted { get; set; }
-    public bool IsReviewed { get; set; }
-    public bool IsBlocked { get; set; } = false;
+
+    // If null, the comment was not reviewed yet
+    public bool? IsRestricted { get; set; }
     public DateTime? DeletedAt { get; set; }
     public CommentUserDTO? User { get; set; }
 
     public ICollection<CommentDTO> Replies { get; set; } = new List<CommentDTO>();
+    public bool IsReviewed => IsRestricted.HasValue;
 }
