+++ Streetcode/Streetcode.BLL/MediatR/Comments/SetCommentRestrictedStatus/SetCommentRestrictedStatusCommand.cs
@@ -1,0 +1,7 @@
+﻿using FluentResults;
+using MediatR;
+using Streetcode.BLL.DTO.Comments;
+
+namespace Streetcode.BLL.MediatR.Comments.SetCommentRestrictedStatus;
+
+public record SetCommentRestrictedStatusCommand(int CommentId, bool IsRestricted) : IRequest<Result<CommentDTO>>;
