+++ Streetcode/Streetcode.BLL/MediatR/Comments/SetCommentRestrictedStatus/SetCommentRestrictedStatusHandler.cs
@@ -7,16 +7,16 @@ using Streetcode.BLL.Resources;
 using Streetcode.BLL.Util.Extensions;
 using Streetcode.DAL.Repositories.Interfaces.Base;
 
-namespace Streetcode.BLL.MediatR.Comments.SetCommentBlockStatus;
+namespace Streetcode.BLL.MediatR.Comments.SetCommentRestrictedStatus;
 
-public class SetCommentBlockStatusHandler
-    : IRequestHandler<SetCommentBlockStatusCommand, Result<CommentDTO>>
+public class SetCommentRestrictedStatusHandler
+    : IRequestHandler<SetCommentRestrictedStatusCommand, Result<CommentDTO>>
 {
     private readonly IRepositoryWrapper _repositoryWrapper;
     private readonly ILoggerService _logger;
     private readonly IMapper _mapper;
 
-    public SetCommentBlockStatusHandler(
+    public SetCommentRestrictedStatusHandler(
         IRepositoryWrapper repositoryWrapper,
         ILoggerService logger,
         IMapper mapper)
@@ -26,7 +26,7 @@ public class SetCommentBlockStatusHandler
         _mapper = mapper;
     }
 
-    public async Task<Result<CommentDTO>> Handle(SetCommentBlockStatusCommand request, CancellationToken cancellationToken)
+    public async Task<Result<CommentDTO>> Handle(SetCommentRestrictedStatusCommand request, CancellationToken cancellationToken)
     {
         var comment = await _repositoryWrapper.CommentRepository
             .GetFirstOrDefaultAsync(c => c.Id == request.CommentId);
@@ -38,13 +38,12 @@ public class SetCommentBlockStatusHandler
             return Result.Fail<CommentDTO>(errorMsg);
         }
 
-        if (comment.IsBlocked == request.Block && comment.IsReviewed)
+        if (comment.IsRestricted == request.IsRestricted && comment.IsReviewed)
         {
             return Result.Ok(_mapper.Map<CommentDTO>(comment));
         }
 
-        comment.IsBlocked = request.Block;
-        comment.IsReviewed = true;
+        comment.IsRestricted = request.IsRestricted;
         comment.UpdatedAt = DateTime.UtcNow;
 
         _repositoryWrapper.CommentRepository.Update(comment);
