+++ Streetcode/Streetcode.DAL/Entities/Comments/CommentContent.cs
@@ -26,9 +26,13 @@ public class CommentContent
     public int? ParentCommentId { get; set; }
     public CommentContent? ParentComment { get; set; }
     public bool IsDeleted { get; set; }
-    public bool IsReviewed { get; set; }
-    public bool IsBlocked { get; set; } = false;
+
+    // If null, the comment was not reviewed yet
+    public bool? IsRestricted { get; set; }
     public DateTime? DeletedAt { get; set; }
 
     public ICollection<CommentContent> Replies { get; set; } = new List<CommentContent>();
+
+    [NotMapped]
+    public bool IsReviewed => IsRestricted.HasValue;
 }
